requirements.txt
alpaca-py==0.36.0
pandas==2.2.3
numpy==2.0.2
scikit-learn==1.5.2
python-dotenv==1.0.1
joblib==1.4.2
fastapi==0.115.8
uvicorn==0.30.6
streamlit==1.39.0
requests==2.32.3
openai>=1.0.0,<2
